+++ dio/lib/src/options.dart
@@ -128,7 +128,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     Map<String, dynamic>? queryParameters,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
-    bool caseSensitiveHeaders = false,
+    bool preserveHeaderCase = false,
     ResponseType? responseType = ResponseType.json,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -147,7 +147,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
           sendTimeout: sendTimeout,
           extra: extra,
           headers: headers,
-          caseSensitiveHeaders: caseSensitiveHeaders,
+          preserveHeaderCase: preserveHeaderCase,
           responseType: responseType,
           contentType: contentType,
           validateStatus: validateStatus,
@@ -175,7 +175,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     Duration? sendTimeout,
     Map<String, Object?>? extra,
     Map<String, Object?>? headers,
-    bool? caseSensitiveHeaders,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -196,7 +196,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
       sendTimeout: sendTimeout ?? this.sendTimeout,
       extra: extra ?? Map.from(this.extra),
       headers: headers ?? Map.from(this.headers),
-      caseSensitiveHeaders: caseSensitiveHeaders ?? this.caseSensitiveHeaders,
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       contentType: contentType ?? this.contentType,
       validateStatus: validateStatus ?? this.validateStatus,
@@ -220,7 +220,7 @@ class Options {
     Duration? receiveTimeout,
     this.extra,
     this.headers,
-    this.caseSensitiveHeaders,
+    this.preserveHeaderCase,
     this.responseType,
     this.contentType,
     this.validateStatus,
@@ -243,7 +243,7 @@ class Options {
     Duration? receiveTimeout,
     Map<String, Object?>? extra,
     Map<String, Object?>? headers,
-    bool? caseSensitiveHeaders,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -280,7 +280,7 @@ class Options {
       receiveTimeout: receiveTimeout ?? this.receiveTimeout,
       extra: extra ?? effectiveExtra,
       headers: headers ?? effectiveHeaders,
-      caseSensitiveHeaders: caseSensitiveHeaders ?? this.caseSensitiveHeaders,
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       contentType: contentType ?? this.contentType,
       validateStatus: validateStatus ?? this.validateStatus,
@@ -329,8 +329,8 @@ class Options {
       baseUrl: baseOpt.baseUrl,
       path: path,
       data: data,
-      caseSensitiveHeaders:
-          caseSensitiveHeaders ?? baseOpt.caseSensitiveHeaders,
+      preserveHeaderCase:
+          preserveHeaderCase ?? baseOpt.preserveHeaderCase,
       sourceStackTrace: sourceStackTrace ?? StackTrace.current,
       connectTimeout: baseOpt.connectTimeout,
       sendTimeout: sendTimeout ?? baseOpt.sendTimeout,
@@ -365,14 +365,14 @@ class Options {
   /// e.g.: `content-type` and `Content-Type` will be treated as the same key.
   Map<String, dynamic>? headers;
 
-  /// Whether the header keys should be case-sensitive.
+  /// Whether the case of header keys should be preserved.
   ///
   /// Defaults to false.
   ///
   /// This option WILL NOT take effect on these circumstances:
   /// - XHR ([HttpRequest]) does not support handling this explicitly.
   /// - The HTTP/2 standard only supports lowercase header keys.
-  bool? caseSensitiveHeaders;
+  bool? preserveHeaderCase;
 
   /// Timeout when sending data.
   ///
@@ -490,7 +490,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
     String? baseUrl,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
-    bool? caseSensitiveHeaders,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -510,7 +510,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
           receiveTimeout: receiveTimeout,
           extra: extra,
           headers: headers,
-          caseSensitiveHeaders: caseSensitiveHeaders,
+          preserveHeaderCase: preserveHeaderCase,
           responseType: responseType,
           contentType: contentType,
           validateStatus: validateStatus,
@@ -543,7 +543,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
     CancelToken? cancelToken,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
-    bool? caseSensitiveHeaders,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -580,7 +580,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
       cancelToken: cancelToken ?? this.cancelToken,
       extra: extra ?? Map.from(this.extra),
       headers: headers ?? Map.from(this.headers),
-      caseSensitiveHeaders: caseSensitiveHeaders ?? this.caseSensitiveHeaders,
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       validateStatus: validateStatus ?? this.validateStatus,
       receiveDataWhenStatusError:
@@ -658,7 +658,7 @@ class _RequestConfig {
     String? method,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
-    bool? caseSensitiveHeaders,
+    bool? preserveHeaderCase,
     String? contentType,
     ListFormat? listFormat,
     bool? followRedirects,
@@ -674,7 +674,7 @@ class _RequestConfig {
         assert(sendTimeout == null || !sendTimeout.isNegative),
         _sendTimeout = sendTimeout,
         method = method ?? 'GET',
-        caseSensitiveHeaders = caseSensitiveHeaders ?? false,
+        preserveHeaderCase = preserveHeaderCase ?? false,
         listFormat = listFormat ?? ListFormat.multi,
         extra = extra ?? {},
         followRedirects = followRedirects ?? true,
@@ -714,7 +714,7 @@ class _RequestConfig {
     }
   }
 
-  late bool caseSensitiveHeaders;
+  late bool preserveHeaderCase;
 
   Duration? get sendTimeout => _sendTimeout;
   Duration? _sendTimeout;