+++ dio/CHANGELOG.md
@@ -7,7 +7,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 - Raise warning for `Map`s other than `Map<String, dynamic>` when encoding request data.
 - Improve exception messages.
-- Allow case-sensitive header keys with the `caseSensitiveHeaders` flag through options.
+- Allow case-sensitive header keys with the `preserveHeaderCase` flag through options.
 
 ## 5.3.3