+++ plugins/native_dio_adapter/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Bump `cronet_http` version.
 - Minimal required Dart version is now 3.1.
 - Minimal required Flutter version is now 3.13.0.
-- Allow case-sensitive header keys with the `caseSensitiveHeaders` flag through options.
+- Allow case-sensitive header keys with the `preserveHeaderCase` flag through options.
 
 ## 1.0.0+2